move AttributeValue to common package
