+++ api/src/main/java/io/opentelemetry/trace/Link.java
@@ -16,6 +16,7 @@
 
 package io.opentelemetry.trace;
 
+import io.opentelemetry.common.AttributeValue;
 import java.util.Map;
 import javax.annotation.concurrent.ThreadSafe;
