+++ exporters/jaeger/src/main/java/io/opentelemetry/exporters/jaeger/Adapter.java
@@ -19,11 +19,11 @@ package io.opentelemetry.exporters.jaeger;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.protobuf.Timestamp;
 import com.google.protobuf.util.Timestamps;
+import io.opentelemetry.common.AttributeValue;
 import io.opentelemetry.exporters.jaeger.proto.api_v2.Model;
 import io.opentelemetry.exporters.otprotocol.TraceProtoUtils;
 import io.opentelemetry.sdk.trace.data.SpanData;
 import io.opentelemetry.sdk.trace.data.SpanData.TimedEvent;
-import io.opentelemetry.trace.AttributeValue;
 import io.opentelemetry.trace.Link;
 import java.util.ArrayList;
 import java.util.Collection;
