+++ opentracing_shim/src/main/java/io/opentelemetry/opentracingshim/SpanShim.java
@@ -16,7 +16,7 @@
 
 package io.opentelemetry.opentracingshim;
 
-import io.opentelemetry.trace.AttributeValue;
+import io.opentelemetry.common.AttributeValue;
 import io.opentelemetry.trace.Status;
 import io.opentracing.Span;
 import io.opentracing.SpanContext;
