+++ sdk/src/main/java/io/opentelemetry/sdk/trace/Sampler.java
@@ -16,7 +16,7 @@
 
 package io.opentelemetry.sdk.trace;
 
-import io.opentelemetry.trace.AttributeValue;
+import io.opentelemetry.common.AttributeValue;
 import io.opentelemetry.trace.Link;
 import io.opentelemetry.trace.Span;
 import io.opentelemetry.trace.SpanContext;
