+++ sdk/src/test/java/io/opentelemetry/sdk/metrics/LongMeasureSdkTest.java
@@ -18,6 +18,7 @@ package io.opentelemetry.sdk.metrics;
 
 import static com.google.common.truth.Truth.assertThat;
 
+import io.opentelemetry.common.AttributeValue;
 import io.opentelemetry.metrics.LongMeasure;
 import io.opentelemetry.metrics.LongMeasure.BoundLongMeasure;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
@@ -30,7 +31,6 @@ import io.opentelemetry.sdk.metrics.data.MetricData.Point;
 import io.opentelemetry.sdk.metrics.data.MetricData.SummaryPoint;
 import io.opentelemetry.sdk.metrics.data.MetricData.ValueAtPercentile;
 import io.opentelemetry.sdk.resources.Resource;
-import io.opentelemetry.trace.AttributeValue;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
