+++ sdk/src/test/java/io/opentelemetry/sdk/trace/TestUtils.java
@@ -16,9 +16,9 @@
 
 package io.opentelemetry.sdk.trace;
 
+import io.opentelemetry.common.AttributeValue;
 import io.opentelemetry.sdk.trace.config.TraceConfig;
 import io.opentelemetry.sdk.trace.data.SpanData;
-import io.opentelemetry.trace.AttributeValue;
 import io.opentelemetry.trace.Span;
 import io.opentelemetry.trace.Span.Kind;
 import io.opentelemetry.trace.SpanId;
