+++ sdk/src/test/java/io/opentelemetry/sdk/trace/data/SpanDataTest.java
@@ -19,9 +19,9 @@ package io.opentelemetry.sdk.trace.data;
 import static com.google.common.truth.Truth.assertThat;
 import static java.util.Collections.emptyList;
 
+import io.opentelemetry.common.AttributeValue;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.trace.data.SpanData.TimedEvent;
-import io.opentelemetry.trace.AttributeValue;
 import io.opentelemetry.trace.Link;
 import io.opentelemetry.trace.Span.Kind;
 import io.opentelemetry.trace.SpanContext;
